refactor: fix typo and class name
