+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/category/id/GetCategoryIdResponse.kt
@@ -3,7 +3,7 @@ package com.binar.secondhand.kel2.data.api.model.seller.category.id
 
 import com.google.gson.annotations.SerializedName
 
-data class getCategoryIdResponse(
+data class GetCategoryIdResponse(
     @SerializedName("createdAt")
     val createdAt: String,
     @SerializedName("id")
