+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/id/get/Category.kt
@@ -1,4 +1,4 @@
-package com.binar.secondhand.kel2.data.api.model.seller.product.id.getProductId
+package com.binar.secondhand.kel2.data.api.model.seller.product.id.get
 
 
 import com.google.gson.annotations.SerializedName
