+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/id/get/GetProductIdResponse.kt
@@ -1,10 +1,14 @@
-package com.binar.secondhand.kel2.data.api.model.seller.banner.id
+package com.binar.secondhand.kel2.data.api.model.seller.product.id.get
 
 
 import com.google.gson.annotations.SerializedName
 
-data class getBannerIdResponse(
-    @SerializedName("createdAt")
+data class GetProductIdResponse(
+    @SerializedName("base_price")
+    val basePrice: Int,
+    @SerializedName("categories")
+    val categories: List<Category>,
+    @SerializedName("created_at")
     val createdAt: String,
     @SerializedName("id")
     val id: Int,
@@ -12,9 +16,11 @@ data class getBannerIdResponse(
     val imageName: String,
     @SerializedName("image_url")
     val imageUrl: String,
+    @SerializedName("location")
+    val location: String,
     @SerializedName("name")
     val name: String,
-    @SerializedName("updatedAt")
+    @SerializedName("updated_at")
     val updatedAt: String,
     @SerializedName("user_id")
     val userId: Int
