+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/id/put/PutProductIdRequest.kt
@@ -1,4 +1,4 @@
-package com.binar.secondhand.kel2.data.api.model.seller.product.id.putProductId
+package com.binar.secondhand.kel2.data.api.model.seller.product.id.put
 
 import com.google.gson.annotations.SerializedName
